refactor: organize page-specific components
