+++ src/pages/Live/components/Live.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import SpectrumBarChart from "../../../components/SpectrumBarChart";
+import SpectrumBarChart from "./SpectrumBarChart";
 import TopicSection from "./TopicSection";
 import NotesBlock from "./NotesBlock";
 import styles from "./SensorDashboard/SensorDashboard.module.css";
