+++ src/pages/Live/components/TopicSection.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import DeviceTable from '../../../components/DeviceTable';
+import DeviceTable from './DeviceTable';
 import styles from './SensorDashboard/SensorDashboard.module.css';
 
 function TopicSection({systemTopics = {}}) {
