+++ src/pages/Reports/components/HistoricalClearLuxChart.jsx
